requirements.txt
cryptography
pysmb
chardet